requirements.txt
pandas>=2.3.2
PyPDF2>=3.0.1
openpyxl>=3.1.5
python-docx>=1.2.0
openai>=1.108.1
jupyter>=1.1.1
notebook>=7.4.5
